[coinone] formated and example add and test package and class name change
