+++ xchange-coinone/pom.xml
@@ -7,7 +7,7 @@
     <parent>
         <groupId>org.knowm.xchange</groupId>
         <artifactId>xchange-parent</artifactId>
-        <version>4.3.5-SNAPSHOT</version>
+        <version>4.3.6-SNAPSHOT</version>
     </parent>
 
     <artifactId>xchange-coinone</artifactId>
