+++ xchange-coinone/src/main/java/org/knowm/xchange/coinone/CoinoneExchange.java
@@ -1,7 +1,6 @@
 package org.knowm.xchange.coinone;
 
 import java.io.IOException;
-
 import org.knowm.xchange.BaseExchange;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeSpecification;
@@ -13,38 +12,37 @@ import org.knowm.xchange.utils.nonce.CurrentTimeNonceFactory;
 import si.mazi.rescu.SynchronizedValueFactory;
 
 public class CoinoneExchange extends BaseExchange implements Exchange {
-	private SynchronizedValueFactory<Long> nonceFactory = new CurrentTimeNonceFactory();
+  private SynchronizedValueFactory<Long> nonceFactory = new CurrentTimeNonceFactory();
 
-	public static enum period {
-		hour,
-		day
-	}
+  public static enum period {
+    hour,
+    day
+  }
 
-	@Override
-	protected void initServices() {
-		this.marketDataService = new CoinoneMarketDataService(this);
-		this.accountService = new CoinoneAccountService(this);
-		this.tradeService = new CoinoneTradeService(this);
-	}
+  @Override
+  protected void initServices() {
+    this.marketDataService = new CoinoneMarketDataService(this);
+    this.accountService = new CoinoneAccountService(this);
+    this.tradeService = new CoinoneTradeService(this);
+  }
 
-	@Override
-	public ExchangeSpecification getDefaultExchangeSpecification() {
-		ExchangeSpecification exchangeSpecification =
-				new ExchangeSpecification(this.getClass().getCanonicalName());
-		exchangeSpecification.setSslUri("https://api.coinone.co.kr");
-		exchangeSpecification.setHost("www.coinone.co.kr");
-		exchangeSpecification.setExchangeName("Coinone");
-		exchangeSpecification.setExchangeDescription("Coinone is a block chain exchange.");
+  @Override
+  public ExchangeSpecification getDefaultExchangeSpecification() {
+    ExchangeSpecification exchangeSpecification =
+        new ExchangeSpecification(this.getClass().getCanonicalName());
+    exchangeSpecification.setSslUri("https://api.coinone.co.kr");
+    exchangeSpecification.setHost("www.coinone.co.kr");
+    exchangeSpecification.setExchangeName("Coinone");
+    exchangeSpecification.setExchangeDescription("Coinone is a block chain exchange.");
 
-		return exchangeSpecification;
-	}
+    return exchangeSpecification;
+  }
 
-	@Override
-	public SynchronizedValueFactory<Long> getNonceFactory() {
-		return nonceFactory;
-	}
+  @Override
+  public SynchronizedValueFactory<Long> getNonceFactory() {
+    return nonceFactory;
+  }
 
-	@Override
-	public void remoteInit() throws IOException, ExchangeException {
-	}
+  @Override
+  public void remoteInit() throws IOException, ExchangeException {}
 }
