+++ xchange-coinone/src/main/java/org/knowm/xchange/coinone/dto/account/CoinoneBalance.java
@@ -4,20 +4,20 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class CoinoneBalance {
 
-	private final double avail;
-	private final double balance;
+  private final double avail;
+  private final double balance;
 
-	public CoinoneBalance(
-			@JsonProperty("avail") String avail, @JsonProperty("balance") String balance) {
-		this.avail = Double.valueOf(avail);
-		this.balance = Double.valueOf(balance);
-	}
+  public CoinoneBalance(
+      @JsonProperty("avail") String avail, @JsonProperty("balance") String balance) {
+    this.avail = Double.valueOf(avail);
+    this.balance = Double.valueOf(balance);
+  }
 
-	public double getAvail() {
-		return avail;
-	}
+  public double getAvail() {
+    return avail;
+  }
 
-	public double getBalance() {
-		return balance;
-	}
+  public double getBalance() {
+    return balance;
+  }
 }
