+++ xchange-coinone/src/main/java/org/knowm/xchange/coinone/dto/marketdata/CoinoneOrderBookData.java
@@ -1,38 +1,35 @@
 package org.knowm.xchange.coinone.dto.marketdata;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-
 import java.math.BigDecimal;
 
-/**
- * @author interwater
- */
+/** @author interwater */
 public class CoinoneOrderBookData {
 
-	private final BigDecimal price;
-	private final BigDecimal qty;
+  private final BigDecimal price;
+  private final BigDecimal qty;
 
-	/**
-	 * @param price
-	 * @param qty
-	 */
-	public CoinoneOrderBookData(
-			@JsonProperty("price") String price, @JsonProperty("qty") String qty) {
-		this.price = new BigDecimal(price);
-		this.qty = new BigDecimal(qty);
-	}
+  /**
+   * @param price
+   * @param qty
+   */
+  public CoinoneOrderBookData(
+      @JsonProperty("price") String price, @JsonProperty("qty") String qty) {
+    this.price = new BigDecimal(price);
+    this.qty = new BigDecimal(qty);
+  }
 
-	public BigDecimal getPrice() {
-		return price;
-	}
+  public BigDecimal getPrice() {
+    return price;
+  }
 
-	public BigDecimal getQty() {
-		return qty;
-	}
+  public BigDecimal getQty() {
+    return qty;
+  }
 
-	@Override
-	public String toString() {
+  @Override
+  public String toString() {
 
-		return "CoinoneOrderBook{" + "price=" + price + ", qty=" + qty + "}";
-	}
+    return "CoinoneOrderBook{" + "price=" + price + ", qty=" + qty + "}";
+  }
 }
