+++ xchange-coinone/src/main/java/org/knowm/xchange/coinone/service/CoinoneAccountService.java
@@ -3,7 +3,6 @@ package org.knowm.xchange.coinone.service;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.util.List;
-
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.coinone.CoinoneAdapters;
 import org.knowm.xchange.coinone.CoinoneExchange;
@@ -19,52 +18,52 @@ import org.knowm.xchange.service.trade.params.WithdrawFundsParams;
 
 public class CoinoneAccountService extends CoinoneAccountServiceRaw implements AccountService {
 
-	private CoinoneExchange coinoneExchange;
+  private CoinoneExchange coinoneExchange;
 
-	/**
-	 * Constructor
-	 *
-	 * @param exchange
-	 */
-	public CoinoneAccountService(Exchange exchange) {
-		super(exchange);
-		this.coinoneExchange = (CoinoneExchange) exchange;
-	}
+  /**
+   * Constructor
+   *
+   * @param exchange
+   */
+  public CoinoneAccountService(Exchange exchange) {
+    super(exchange);
+    this.coinoneExchange = (CoinoneExchange) exchange;
+  }
 
-	@Override
-	public AccountInfo getAccountInfo()
-			throws ExchangeException, NotAvailableFromExchangeException,
-			NotYetImplementedForExchangeException, IOException {
-		return new AccountInfo(CoinoneAdapters.adaptWallet(super.getWallet()));
-	}
+  @Override
+  public AccountInfo getAccountInfo()
+      throws ExchangeException, NotAvailableFromExchangeException,
+          NotYetImplementedForExchangeException, IOException {
+    return new AccountInfo(CoinoneAdapters.adaptWallet(super.getWallet()));
+  }
 
-	@Override
-	public String withdrawFunds(Currency currency, BigDecimal amount, String address)
-			throws IOException {
-		return null;
-	}
+  @Override
+  public String withdrawFunds(Currency currency, BigDecimal amount, String address)
+      throws IOException {
+    return null;
+  }
 
-	@Override
-	public String withdrawFunds(WithdrawFundsParams withdrawFundsParams) throws IOException {
-		return null;
-	}
+  @Override
+  public String withdrawFunds(WithdrawFundsParams withdrawFundsParams) throws IOException {
+    return null;
+  }
 
-	@Override
-	public String requestDepositAddress(Currency currency, String... args)
-			throws ExchangeException, NotAvailableFromExchangeException,
-			NotYetImplementedForExchangeException, IOException {
-		throw new NotYetImplementedForExchangeException();
-	}
+  @Override
+  public String requestDepositAddress(Currency currency, String... args)
+      throws ExchangeException, NotAvailableFromExchangeException,
+          NotYetImplementedForExchangeException, IOException {
+    throw new NotYetImplementedForExchangeException();
+  }
 
-	@Override
-	public TradeHistoryParams createFundingHistoryParams() {
-		throw new NotYetImplementedForExchangeException();
-	}
+  @Override
+  public TradeHistoryParams createFundingHistoryParams() {
+    throw new NotYetImplementedForExchangeException();
+  }
 
-	@Override
-	public List<FundingRecord> getFundingHistory(TradeHistoryParams params)
-			throws ExchangeException, NotAvailableFromExchangeException,
-			NotYetImplementedForExchangeException, IOException {
-		throw new NotYetImplementedForExchangeException();
-	}
+  @Override
+  public List<FundingRecord> getFundingHistory(TradeHistoryParams params)
+      throws ExchangeException, NotAvailableFromExchangeException,
+          NotYetImplementedForExchangeException, IOException {
+    throw new NotYetImplementedForExchangeException();
+  }
 }
