+++ xchange-coinone/src/main/java/org/knowm/xchange/coinone/service/CoinoneAccountServiceRaw.java
@@ -2,7 +2,6 @@ package org.knowm.xchange.coinone.service;
 
 import java.io.IOException;
 import java.math.BigDecimal;
-
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.coinone.dto.CoinoneException;
 import org.knowm.xchange.coinone.dto.account.*;
@@ -10,36 +9,36 @@ import org.knowm.xchange.currency.Currency;
 
 public class CoinoneAccountServiceRaw extends CoinoneBaseService {
 
-	/**
-	 * Constructor
-	 *
-	 * @param exchange
-	 */
-	public CoinoneAccountServiceRaw(Exchange exchange) {
+  /**
+   * Constructor
+   *
+   * @param exchange
+   */
+  public CoinoneAccountServiceRaw(Exchange exchange) {
 
-		super(exchange);
-	}
+    super(exchange);
+  }
 
-	public CoinoneBalancesResponse getWallet() throws CoinoneException, IOException {
-		CoinoneBalancesRequest request =
-				new CoinoneBalancesRequest(apiKey, exchange.getNonceFactory().createValue());
-		CoinoneBalancesResponse response = coinone.getWallet(payloadCreator, signatureCreator, request);
-		return response;
-	}
+  public CoinoneBalancesResponse getWallet() throws CoinoneException, IOException {
+    CoinoneBalancesRequest request =
+        new CoinoneBalancesRequest(apiKey, exchange.getNonceFactory().createValue());
+    CoinoneBalancesResponse response = coinone.getWallet(payloadCreator, signatureCreator, request);
+    return response;
+  }
 
-	public CoinoneWithdrawResponse withdrawFund(
-			Currency currency, BigDecimal amount, String address, String authNumber)
-			throws CoinoneException, IOException {
-		CoinoneWithdrawRequest request =
-				new CoinoneWithdrawRequest(
-						apiKey,
-						exchange.getNonceFactory().createValue(),
-						currency.getSymbol().toLowerCase(),
-						authNumber,
-						amount.doubleValue(),
-						address);
-		CoinoneWithdrawResponse response =
-				coinone.withdrawFund(payloadCreator, signatureCreator, request);
-		return response;
-	}
+  public CoinoneWithdrawResponse withdrawFund(
+      Currency currency, BigDecimal amount, String address, String authNumber)
+      throws CoinoneException, IOException {
+    CoinoneWithdrawRequest request =
+        new CoinoneWithdrawRequest(
+            apiKey,
+            exchange.getNonceFactory().createValue(),
+            currency.getSymbol().toLowerCase(),
+            authNumber,
+            amount.doubleValue(),
+            address);
+    CoinoneWithdrawResponse response =
+        coinone.withdrawFund(payloadCreator, signatureCreator, request);
+    return response;
+  }
 }
