+++ xchange-coinone/src/main/java/org/knowm/xchange/coinone/service/CoinoneMarketDataService.java
@@ -1,7 +1,6 @@
 package org.knowm.xchange.coinone.service;
 
 import java.io.IOException;
-
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.coinone.CoinoneAdapters;
 import org.knowm.xchange.coinone.CoinoneExchange;
@@ -14,45 +13,43 @@ import org.knowm.xchange.exceptions.NotAvailableFromExchangeException;
 import org.knowm.xchange.exceptions.NotYetImplementedForExchangeException;
 import org.knowm.xchange.service.marketdata.MarketDataService;
 
-/**
- * @author interwater
- */
+/** @author interwater */
 public class CoinoneMarketDataService extends CoinoneMarketDataServiceRaw
-		implements MarketDataService {
+    implements MarketDataService {
 
-	/**
-	 * Constructor
-	 *
-	 * @param exchange
-	 */
-	public CoinoneMarketDataService(Exchange exchange) {
+  /**
+   * Constructor
+   *
+   * @param exchange
+   */
+  public CoinoneMarketDataService(Exchange exchange) {
 
-		super(exchange);
-	}
+    super(exchange);
+  }
 
-	@Override
-	public Ticker getTicker(CurrencyPair currencyPair, Object... args) throws IOException {
-		return CoinoneAdapters.adaptTicker(super.getTicker(currencyPair));
-	}
+  @Override
+  public Ticker getTicker(CurrencyPair currencyPair, Object... args) throws IOException {
+    return CoinoneAdapters.adaptTicker(super.getTicker(currencyPair));
+  }
 
-	@Override
-	public OrderBook getOrderBook(CurrencyPair currencyPair, Object... args)
-			throws ExchangeException, NotAvailableFromExchangeException,
-			NotYetImplementedForExchangeException, IOException {
-		return CoinoneAdapters.adaptOrderBook(getCoinoneOrderBook(currencyPair), currencyPair);
-	}
+  @Override
+  public OrderBook getOrderBook(CurrencyPair currencyPair, Object... args)
+      throws ExchangeException, NotAvailableFromExchangeException,
+          NotYetImplementedForExchangeException, IOException {
+    return CoinoneAdapters.adaptOrderBook(getCoinoneOrderBook(currencyPair), currencyPair);
+  }
 
-	@Override
-	public Trades getTrades(CurrencyPair currencyPair, Object... args)
-			throws ExchangeException, NotAvailableFromExchangeException,
-			NotYetImplementedForExchangeException, IOException {
-		String period = "hour";
-		if (args[0] != null) {
-			try {
-				period = CoinoneExchange.period.valueOf(args[0].toString()).name();
-			} catch (IllegalArgumentException e) {
-			}
-		}
-		return CoinoneAdapters.adaptTrades(super.getTrades(currencyPair, period), currencyPair);
-	}
+  @Override
+  public Trades getTrades(CurrencyPair currencyPair, Object... args)
+      throws ExchangeException, NotAvailableFromExchangeException,
+          NotYetImplementedForExchangeException, IOException {
+    String period = "hour";
+    if (args[0] != null) {
+      try {
+        period = CoinoneExchange.period.valueOf(args[0].toString()).name();
+      } catch (IllegalArgumentException e) {
+      }
+    }
+    return CoinoneAdapters.adaptTrades(super.getTrades(currencyPair, period), currencyPair);
+  }
 }
