+++ xchange-coinone/src/main/java/org/knowm/xchange/coinone/service/CoinoneTradeService.java
@@ -2,7 +2,6 @@ package org.knowm.xchange.coinone.service;
 
 import java.io.IOException;
 import java.util.Collection;
-
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.coinone.dto.CoinoneException;
 import org.knowm.xchange.coinone.dto.trade.CoinoneTradeCancelRequest;
@@ -18,92 +17,90 @@ import org.knowm.xchange.service.trade.params.TradeHistoryParams;
 import org.knowm.xchange.service.trade.params.orders.OpenOrdersParams;
 import org.knowm.xchange.service.trade.params.orders.OrderQueryParams;
 
-/**
- * @author interwater
- */
+/** @author interwater */
 public class CoinoneTradeService extends CoinoneTradeServiceRaw implements TradeService {
-	/**
-	 * Constructor
-	 *
-	 * @param exchange
-	 */
-	public CoinoneTradeService(Exchange exchange) {
+  /**
+   * Constructor
+   *
+   * @param exchange
+   */
+  public CoinoneTradeService(Exchange exchange) {
 
-		super(exchange);
-	}
+    super(exchange);
+  }
 
-	@Override
-	public OpenOrders getOpenOrders() throws IOException {
-		return getOpenOrders(createOpenOrdersParams());
-	}
+  @Override
+  public OpenOrders getOpenOrders() throws IOException {
+    return getOpenOrders(createOpenOrdersParams());
+  }
 
-	@Override
-	public OpenOrders getOpenOrders(OpenOrdersParams params)
-			throws ExchangeException, NotAvailableFromExchangeException,
-			NotYetImplementedForExchangeException, IOException {
-		throw new NotYetImplementedForExchangeException();
-	}
+  @Override
+  public OpenOrders getOpenOrders(OpenOrdersParams params)
+      throws ExchangeException, NotAvailableFromExchangeException,
+          NotYetImplementedForExchangeException, IOException {
+    throw new NotYetImplementedForExchangeException();
+  }
 
-	@Override
-	public String placeMarketOrder(MarketOrder marketOrder)
-			throws IOException, NotYetImplementedForExchangeException {
-		throw new NotYetImplementedForExchangeException();
-	}
+  @Override
+  public String placeMarketOrder(MarketOrder marketOrder)
+      throws IOException, NotYetImplementedForExchangeException {
+    throw new NotYetImplementedForExchangeException();
+  }
 
-	@Override
-	public String placeLimitOrder(LimitOrder limitOrder)
-			throws IOException, NotAvailableFromExchangeException {
-		CoinoneTradeResponse response = super.placeLimitOrderRaw(limitOrder);
-		if (!response.getErrorCode().equals("0")) {
-			throw new CoinoneException(CoinoneException.resMsgMap.get(response.getErrorCode()));
-		}
-		return response.getOrderId();
-	}
+  @Override
+  public String placeLimitOrder(LimitOrder limitOrder)
+      throws IOException, NotAvailableFromExchangeException {
+    CoinoneTradeResponse response = super.placeLimitOrderRaw(limitOrder);
+    if (!response.getErrorCode().equals("0")) {
+      throw new CoinoneException(CoinoneException.resMsgMap.get(response.getErrorCode()));
+    }
+    return response.getOrderId();
+  }
 
-	@Override
-	public String placeStopOrder(StopOrder stopOrder) throws IOException {
-		return null;
-	}
+  @Override
+  public String placeStopOrder(StopOrder stopOrder) throws IOException {
+    return null;
+  }
 
-	public boolean cancelOrder(String orderId) throws IOException {
-		throw new NotYetImplementedForExchangeException();
-	}
+  public boolean cancelOrder(String orderId) throws IOException {
+    throw new NotYetImplementedForExchangeException();
+  }
 
-	@Override
-	public UserTrades getTradeHistory(TradeHistoryParams params) throws IOException {
-		throw new NotYetImplementedForExchangeException();
-	}
+  @Override
+  public UserTrades getTradeHistory(TradeHistoryParams params) throws IOException {
+    throw new NotYetImplementedForExchangeException();
+  }
 
-	@Override
-	public TradeHistoryParams createTradeHistoryParams() {
-		throw new NotYetImplementedForExchangeException();
-	}
+  @Override
+  public TradeHistoryParams createTradeHistoryParams() {
+    throw new NotYetImplementedForExchangeException();
+  }
 
-	@Override
-	public OpenOrdersParams createOpenOrdersParams() {
-		throw new NotYetImplementedForExchangeException();
-	}
+  @Override
+  public OpenOrdersParams createOpenOrdersParams() {
+    throw new NotYetImplementedForExchangeException();
+  }
 
-	@Override
-	public Collection<Order> getOrder(String... orderIds)
-			throws ExchangeException, NotAvailableFromExchangeException,
-			NotYetImplementedForExchangeException, IOException {
-		throw new NotYetImplementedForExchangeException();
-	}
+  @Override
+  public Collection<Order> getOrder(String... orderIds)
+      throws ExchangeException, NotAvailableFromExchangeException,
+          NotYetImplementedForExchangeException, IOException {
+    throw new NotYetImplementedForExchangeException();
+  }
 
-	@Override
-	public Collection<Order> getOrder(OrderQueryParams... orderQueryParams) throws IOException {
-		return null;
-	}
+  @Override
+  public Collection<Order> getOrder(OrderQueryParams... orderQueryParams) throws IOException {
+    return null;
+  }
 
-	@Override
-	public boolean cancelOrder(CancelOrderParams orderParams)
-			throws ExchangeException, NotAvailableFromExchangeException,
-			NotYetImplementedForExchangeException, IOException {
-		CoinoneTradeResponse response = super.cancerOrder((CoinoneTradeCancelRequest) orderParams);
-		if (!response.getErrorCode().equals("0")) {
-			throw new CoinoneException(CoinoneException.resMsgMap.get(response.getErrorCode()));
-		}
-		return true;
-	}
+  @Override
+  public boolean cancelOrder(CancelOrderParams orderParams)
+      throws ExchangeException, NotAvailableFromExchangeException,
+          NotYetImplementedForExchangeException, IOException {
+    CoinoneTradeResponse response = super.cancerOrder((CoinoneTradeCancelRequest) orderParams);
+    if (!response.getErrorCode().equals("0")) {
+      throw new CoinoneException(CoinoneException.resMsgMap.get(response.getErrorCode()));
+    }
+    return true;
+  }
 }
