+++ xchange-coinone/src/test/java/org/knowm/xchange/coinone/service/marketdata/OrderBookIntegration.java
@@ -1,4 +1,4 @@
-package org.knowm.xchange.coinone.service;
+package org.knowm.xchange.coinone.service.marketdata;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,16 +8,16 @@ import org.knowm.xchange.ExchangeFactory;
 import org.knowm.xchange.ExchangeSpecification;
 import org.knowm.xchange.coinone.CoinoneExchange;
 import org.knowm.xchange.currency.CurrencyPair;
-import org.knowm.xchange.dto.marketdata.Ticker;
+import org.knowm.xchange.dto.marketdata.OrderBook;
 import org.knowm.xchange.service.marketdata.MarketDataService;
 
-public class TickerTest {
+public class OrderBookIntegration {
   @Test
-  public void tickerTest() throws Exception {
+  public void orderBookTest() throws Exception {
     ExchangeSpecification exSpec = new ExchangeSpecification(CoinoneExchange.class);
     Exchange exchange = ExchangeFactory.INSTANCE.createExchange(exSpec);
     MarketDataService marketDataService = exchange.getMarketDataService();
-    Ticker ticker = marketDataService.getTicker(CurrencyPair.BTC_KRW);
-    assertThat(ticker).isNotNull();
+    OrderBook orderBook = marketDataService.getOrderBook(CurrencyPair.ETH_BTC);
+    assertThat(orderBook).isNotNull();
   }
 }
