+++ xchange-coinone/src/test/java/org/knowm/xchange/coinone/service/marketdata/TickerIntegration.java
@@ -1,0 +1,23 @@
+package org.knowm.xchange.coinone.service.marketdata;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.Test;
+import org.knowm.xchange.Exchange;
+import org.knowm.xchange.ExchangeFactory;
+import org.knowm.xchange.ExchangeSpecification;
+import org.knowm.xchange.coinone.CoinoneExchange;
+import org.knowm.xchange.currency.CurrencyPair;
+import org.knowm.xchange.dto.marketdata.Ticker;
+import org.knowm.xchange.service.marketdata.MarketDataService;
+
+public class TickerIntegration {
+  @Test
+  public void tickerTest() throws Exception {
+    ExchangeSpecification exSpec = new ExchangeSpecification(CoinoneExchange.class);
+    Exchange exchange = ExchangeFactory.INSTANCE.createExchange(exSpec);
+    MarketDataService marketDataService = exchange.getMarketDataService();
+    Ticker ticker = marketDataService.getTicker(CurrencyPair.BTC_KRW);
+    assertThat(ticker).isNotNull();
+  }
+}
