+++ xchange-examples/pom.xml
@@ -297,6 +297,11 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>xchange-coinone</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>xchange-empoex</artifactId>
             <version>${project.version}</version>
         </dependency>
